Move test_no_shadowed_builtins out to functional tests
